removed reliance on MovieImport.numberOfImages()
In my test, MovieImport.numberOfImages returned 0 when it should have returned 155. I added code to use  _tempFolder.content.length to count the number of files created for the import, as a hack. 
After this modification, the importQT function executes without errors. I do note though that every time navigate to a frame, the message log says that "path/to/folder/for/element/frameFileName-small.tga" doesn't exist. It then seems to spend a moment to make that file. Should the import process makes these .tga previews for Harmony, or just leave to do itself?