Fix scene.length setter
Setting a scene's length to shorter than the current length produced no result when I tested it in H20. I think it's because `frame.remove()` expects a positive value for the number of frames to remove.
I rearranged it so that it starts countign frames to remove at `scene.length - difference`, then removed `difference` frames.
I also removed the -1 offset present in the original. Not sure what it was doing, but maybe I don't realise the importance. If we used `ripple_markers` or `extend_exposures` maybe it would do something?