Update filebrowser functions to return oFiles or oFolders
I noticed that browseForFolder was returning a string, which was unexpected because oH tends to use oFolders and oFiles wherever possible, rather than raw paths.

I'm also unclear what `for (i in _files)` would do in the case of a single string, rather than a string array, so I added an explicit test for an array return.

Not sure if the replace() filters are necessary since oFile parses paths. Also not sure if this kind of change is ok, since it could possibly break things that rely on string returns? But oFile/oFolders coerce into their string paths anyway.