Added choose File/Folder funcs
Added variants of browseForFile and browseForFolder function that return oFile and oFolder objects.

They work as is, but I wonder if it would be better to have one function with flags for acceptMultiple and some kind of file/folder/both switch? There's no way to select multiple folders as is, although I couldn't find a way to do this trying to discern the Qt docs.