Initial concept for scaling imported images to the scene's field chart
Added new param to end of func for compatibility, but maybe it could be a new value for 'alignment' `PROJECT_RESOLUTION` which would be `ASIS` + the scaling logic?

I'm not sure if my logic for scaling the image is short-sighted in some way, since I haven't worked with scenes with non-default field setups. Since oH historically would have defaulted to 12 fields, maybe this logic should as well, in case anyone was relying on that behaviour.

Another concept would be to have an extra param that could be any of `FIT_Y`, `FIT_X`, `PROJECT_RESOLUTION` or a scale factor for manual adjustment.

The benefit I see for this method of setting fields is that the user can set the scale to 1 and the image will go back to the 'correct' size. The drawback is that it's different to how the GUI import process works, so images brought in via script and via user would mysteriously not work with the same pegs