fields adjustment
A scene with a field grid described as "12 Fields" actually has 24x24 units. The `element.add()` function takes a 'field guide' arg, which corresponds not to the number of field units, but to an integer for the 12 in 12-field (which means 24). So all my imports were off by a factor of 2.

I have tested float values and it behaves as expected, so it *should* handle weird field guides if people set up od numbers or something