used __file__ from import-time
when running a script in the sandbox, __file__ is undefined, so trying to look at $ in the debugger errors out when it tries to generate the oH directory. If we use the __file__ it stored when it was required in, we can avoid this error. $.directory is supposed to refer to the oH dir, so I can't think of why this would be a problem