broken link to $ in oPreference
For some reason, `this.$` is undefined for this line, however `this.prototype.$` is. With the debug console I tried to set `this.$ = this.prototype.$`, and it didn't work, it stayed undefined. Then I tried using an intermediary constant, which did work.

Referencing `exports` bypasses the issue.

Is this a different case because we're not instantiating an oPreferences object, so the prototype isn't linked in the same way?