broken refs to this
two fixes i needed to make for some existing scripts to work with requiring oH in the context of a toolbar.

ln 2514:
seems like in the `map()` scope, `this` isn't the oBackdrop. Hack solution to save a var ref to `this` as `_this` which the map can see

ln 2883:
missed bare ref to oPoint replaced with this.$.oPoint